scraping-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
selenium==4.15.2
webdriver-manager==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
pillow==10.1.0
python-multipart==0.0.6
aiofiles==23.2.1
beautifulsoup4==4.12.2
pandas==2.1.3
openpyxl==3.1.2
openai>=1.30.0